Add ValueAddedTaxCalculator service and related configurations
This commit introduces a new service, `ValueAddedTaxCalculator`, along with its corresponding project file, models, and extensions. It essentially calculates various quantities related to purchase such as NetAmount, GrossAmount, and VatAmount based on the inputs given. Additionally, launch settings and Swagger packages for API documentation are added. The project also includes an HTTP file for testing endpoints and a `.gitignore` file to ignore unnecessary files and folders.