Add API endpoint documentation to README
The README file has been updated to include documentation about the newly added REST API endpoint, `/calculate-purchase`, for accessing the calculation service. In addition, the README file has been added to the project's Solution Items. The format of the code usage example in the README file is also improved for readability.